strategic-decision-engine/requirements.txt
fastapi
streamlit
langchain
langchain-nomic
langchain-community
langchain-pinecone
langchain_groq
ragas
pinecone-client
redis
uvicorn
python-dotenv
psycopg2-binary
pymongo 